refactor(render): split the dashboard's sections from its Worker table
The file-size ratchet caught `dashboard.ts` at 893 lines, over its 800-line
threshold, and its message is explicit that the baseline records debt predating
the ratchet and never a new file. So this splits by DOMAIN rather than by size:
`dashboard-sections.ts` answers "what is this host watching, and how has it been
going", while what remains answers "what is each Worker doing right now".

The sparkline goes with the first question, and it is EXPORTED — two other
modules in this repo grew their own copy of the same eight glyphs, and a shared
one is the point.

dashboard.ts: 893 → 751 lines.
